requirements.txt
# Core AI & Data
google-genai              # The new Google GenAI SDK
altair                    # Visualization (used by Streamlit)
fpdf2                     # PDF Generation
pypdf                     # PDF Merging/Reading
pymongo                   # MongoDB Database connection

# Web & Utilities
streamlit                 # The Web App Framework
watchdog                  # For Streamlit auto-reload
python-dotenv             # To read .env files
requests                  # HTTP Requests

# Authentication
Authlib                   # For Google OAuth
google-auth               # Google Auth Utilities
uharfbuzz                 # CRITICAL: For Marathi Font support in FPDF2